backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
requests==2.31.0
python-decouple==3.8
cryptography==42.0.8
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.124.0
google-generativeai==0.3.2